Refactor ProtocolInfo to separate utility class.
